Change Loop IO to bool toggle
Most puzzle genres don't actually need the concept of inside / outside the loop, only that something is on the loop, which can be done through something like `sym.is_loop(sg.grid[p])`. Moving the loop IO constrainer to a boolean toggle avoids the somewhat expensive construction when not needed.